refactor spring namespace example.
