+++ sharding-jdbc-example/spring-namespace-example/spring-namespace-jpa-example/pom.xml
@@ -28,10 +28,5 @@
             <artifactId>spring-orm</artifactId>
             <version>${spring-framework.version}</version>
         </dependency>
-        <dependency>
-            <groupId>io.shardingsphere</groupId>
-            <artifactId>spring-namespace-base</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
 </project>
