+++ sharding-jdbc-example/spring-namespace-example/spring-namespace-jpa-example/src/main/resources/META-INF/orche/zookeeper/local/shardingDatabases.xml
@@ -13,8 +13,8 @@
                         http://www.springframework.org/schema/context/spring-context.xsd
                         http://www.springframework.org/schema/tx
                         http://www.springframework.org/schema/tx/spring-tx.xsd">
-    <context:annotation-config />
     <import resource="classpath*:META-INF/orche/zookeeper/regCenter.xml" />
+    <context:annotation-config />
     <context:component-scan base-package="io.shardingsphere.example.spring.namespace.jpa"/>
     
     <bean id="entityManagerFactory" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
