+++ sharding-jdbc-example/spring-namespace-example/spring-namespace-jpa-example/src/main/resources/META-INF/orche/zookeeper/local/shardingMasterSlave.xml
@@ -16,8 +16,8 @@
                         http://shardingsphere.io/schema/shardingsphere/orchestration/sharding/sharding.xsd
                         http://shardingsphere.io/schema/shardingsphere/masterslave
                         http://shardingsphere.io/schema/shardingsphere/masterslave/master-slave.xsd">
-    <context:annotation-config />
     <import resource="classpath*:META-INF/orche/zookeeper/regCenter.xml" />
+    <context:annotation-config />
     <context:component-scan base-package="io.shardingsphere.example.spring.namespace.jpa"/>
     
     <bean id="entityManagerFactory" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
