+++ sharding-jdbc-example/spring-namespace-example/spring-namespace-mybatis-example/pom.xml
@@ -23,10 +23,5 @@
             <artifactId>spring-jdbc</artifactId>
             <version>${spring-framework.version}</version>
         </dependency>
-        <dependency>
-            <groupId>io.shardingsphere</groupId>
-            <artifactId>spring-namespace-base</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
 </project>
